# spec/schema/@typespec/openapi3/openapi.yaml
openapi: 3.0.0
info:
  title: Auth
  version: 0.0.0
...
paths:
...
  /authorize:
    get:
      operationId: authorize
      summary: Authorization Request
      parameters:
        - name: response_type
          in: query
          required: true
          schema:
            type: string
          explode: false
        - name: client_id
          in: query
          required: true
          schema:
            type: string
          explode: false
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
          explode: false
        - name: scope
          in: query
          required: true
          schema:
            type: string
          explode: false
        - name: state
          in: query
          required: false
          schema:
            type: string
          explode: false
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        '302':
          description: Redirection
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  error_description:
                    type: string
                  state:
                    type: string
                required:
                  - error
                  - error_description
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  error_description:
                    type: string
                  state:
                    type: string
                required:
                  - error
                  - error_description
    post:
      operationId: postAuthorize
      summary: Authorization Request
      parameters: []
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        '302':
          description: Redirection
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  error_description:
                    type: string
                  state:
                    type: string
                required:
                  - error
                  - error_description
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  error_description:
                    type: string
                  state:
                    type: string
                required:
                  - error
                  - error_description
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  error_description:
                    type: string
                  state:
                    type: string
                required:
                  - error
                  - error_description
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                response_type:
                  type: string
                client_id:
                  type: string
                redirect_uri:
                  type: string
                scope:
                  type: string
                state:
                  type: string
              required:
                - response_type
                - client_id
                - redirect_uri
                - scope
...
servers:
  - url: https://auth.piny940.com/api/v1
    description: Auth Server
    variables: {}